Automatic repository split for publishing on packagist (#171)
Adds the setup necessary for distributing php-toolkit packages via
composer:

https://packagist.org/packages/wp-php-toolkit/

### Data flows

Composer requires one git repository per package, and the package name
depends on the repository owner and name. This PR wires the publishing
pipeline to a new `wp-php-toolkit` organization. It splits the directory
tree at `components/*`, creates a new repo for every package that
doesn't have one yet, and pushes the code.

### GitHub organization

The packages are published under . I don't have access to create new
repos in the `WordPress` org or publish composer packages under the
`WordPress` org. I'm happy to rewire the publishing pipeline to use the
WordPress org if there's interest in that, but I'm starting with
`wp-php-toolkit` to keep things moving.

### Upcoming work

* Ensure git file tree split is done before publishing new packages on
composer
* Run smoke tests after pushing to git but before publishing new
packages on composer
* Add README and at least rudimentary documentation to each component

---------

Co-authored-by: Adam Zielinski <cloudnik@MacBook-Air-Adam.local>